requirements.txt
asgiref==3.2.3
autopep8==1.5
backports.csv==1.0.7
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.3
django-admin-rangefilter==0.5.0
django-advanced-filters==1.1.1
django-braces==1.14.0
django-common-helpers==0.9.2
django-cors-headers==3.1.0
django-cron==0.5.1
django-filter==2.2.0
django-import-export==1.2.0
django-rest-swagger==2.2.0
django-silk==4.0.0
django-storages==1.7.2
django-unique-upload==0.2.1
djangorestframework==3.10.3
docutils==0.15.2
et-xmlfile==1.0.1
gprof2dot==2019.11.30
gunicorn==20.1.0
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.4
MarkupSafe==1.1.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==3.0.0
psycopg2-binary==2.9.1
pycodestyle==2.5.0
Pygments==2.5.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.3.2
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.1
urllib3==1.25.6
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
